requirements.txt
arabic_reshaper==3.0.0
fpdf==1.7.2
fpdf2==2.8.3
pandas==2.3.0
pdfplumber==0.11.7
streamlit==1.46.0
python-bidi==0.6.6
openpyxl==3.1.5

